I want to build a Time-based One-Time Password (TOTP) app similar to Google Authenticator. The app will be called authr.
The app should allow users to enter acccount names and seed keys for TOTP generation.

The project will have libraries for business logic and accounts management.

There will be a number of UIs starting with a simple command line interface:

authr list // list accounts
authr add // add an account
authr remove // remove an account
authr show <account name> [--seed] // show an account name and the current TOTP, or the seed key if --seed is specified

The second UI will be curses-based like the top command, invoked wihen no command line options are specified.
it will show all accounts and their current TOTPs along with a timer.
It supports typing to filter the list of accounts. 
the ESC key will exit.

For the first pass the storage for account info can be in a simple database or file. 
The file location should be an OS appropriate user specific location such as `~/.config/`.  

Future requirements will include encrypting the storage and eventually the ability to save in the cloud, the details of which are to be determined later.

The app will be written in Rust language. It should use open source crates for curses, TOTP generation, and any other appropriate options.